fix: colors ✔

+++ DOCKER/scripts/sites.sh
@@ -28,30 +28,36 @@ newsite()
          subdir="build";
       ;;
    esac
-   h1 $YELLOW "${typefile^^} :: https://${sites}.${COMPOSE_PROJECT_NAME,,}.local" '|' " "
 
-   mkdir -p www/sites/${sites}/${subdir}
-   mkcert -install
-
-   cp -v DOCKER/examples/examplesite-${typefile}-local.conf DOCKER/images/nginx/sites/${typefile}-${sites}_${COMPOSE_PROJECT_NAME,,}_local.conf
-   sed -i "s/examplesite/${sites}/g" DOCKER/images/nginx/sites/${typefile}-${sites}_${COMPOSE_PROJECT_NAME,,}_local.conf
-   sed -i "s/COMPOSE_PROJECT_NAME/${COMPOSE_PROJECT_NAME,,}/g" DOCKER/images/nginx/sites/${typefile}-${sites}_${COMPOSE_PROJECT_NAME,,}_local.conf
-
-   echo -e "127.0.0.1\t\t${sites}.${COMPOSE_PROJECT_NAME,,}.local www.${sites}.${COMPOSE_PROJECT_NAME,,}.local" | sudo tee -a /etc/hosts
-
-   cd DOCKER/certs
-
-   mkcert ${sites}.${COMPOSE_PROJECT_NAME,,}.local www.${sites}.${COMPOSE_PROJECT_NAME,,}.local
-   mv -v ${sites}.${COMPOSE_PROJECT_NAME,,}.local*-key.pem certs_${sites}_${COMPOSE_PROJECT_NAME,,}_local-key.pem 
-   mv -v ${sites}.${COMPOSE_PROJECT_NAME,,}.local*.pem certs_${sites}_${COMPOSE_PROJECT_NAME,,}_local.pem 
-
-   cd ../..
-   more DOCKER/images/nginx/sites/${typefile}-${sites}_${COMPOSE_PROJECT_NAME,,}_local.conf | grep server_name | head -1
-   if [ ! -f "${COMPOSE_PROJECT_NAME,,}.md" ]; then
-      echo "# SITIOS " > ${COMPOSE_PROJECT_NAME,,}.md
-   fi
-   echo -e " *  [${sites^^}](https://${sites}.${COMPOSE_PROJECT_NAME,,}.local) :: ${typefile^^}" >> ${COMPOSE_PROJECT_NAME,,}.md
-   docker restart homelab-webserver
-   h1  $LIGTH_CYAN "  done ... (${sites}.${COMPOSE_PROJECT_NAME,,}.local) " "" "."
+   R1 $YELLOW "${typefile^^} :: https://${sites}.${COMPOSE_PROJECT_NAME,,}.local" $WHITE "⛁" "."
    ln
+   if [ ! -f "DOCKER/images/nginx/sites/${typefile}-${sites}_${COMPOSE_PROJECT_NAME,,}_local.conf" ]; then
+
+      mkdir -p www/sites/${sites}/${subdir}
+      mkcert -install
+
+      cp -v DOCKER/examples/examplesite-${typefile}-local.conf DOCKER/images/nginx/sites/${typefile}-${sites}_${COMPOSE_PROJECT_NAME,,}_local.conf
+      sed -i "s/examplesite/${sites}/g" DOCKER/images/nginx/sites/${typefile}-${sites}_${COMPOSE_PROJECT_NAME,,}_local.conf
+      sed -i "s/COMPOSE_PROJECT_NAME/${COMPOSE_PROJECT_NAME,,}/g" DOCKER/images/nginx/sites/${typefile}-${sites}_${COMPOSE_PROJECT_NAME,,}_local.conf
+
+      echo -e "127.0.0.1\t\t${sites}.${COMPOSE_PROJECT_NAME,,}.local www.${sites}.${COMPOSE_PROJECT_NAME,,}.local" | sudo tee -a /etc/hosts
+
+      cd DOCKER/certs
+
+      mkcert ${sites}.${COMPOSE_PROJECT_NAME,,}.local www.${sites}.${COMPOSE_PROJECT_NAME,,}.local
+      mv -v ${sites}.${COMPOSE_PROJECT_NAME,,}.local*-key.pem certs_${sites}_${COMPOSE_PROJECT_NAME,,}_local-key.pem 
+      mv -v ${sites}.${COMPOSE_PROJECT_NAME,,}.local*.pem certs_${sites}_${COMPOSE_PROJECT_NAME,,}_local.pem 
+
+      cd ../..
+      more DOCKER/images/nginx/sites/${typefile}-${sites}_${COMPOSE_PROJECT_NAME,,}_local.conf | grep server_name | head -1
+      if [ ! -f "${COMPOSE_PROJECT_NAME,,}.md" ]; then
+         echo "# SITIOS " > ${COMPOSE_PROJECT_NAME,,}.md
+      fi
+      echo -e " *  [${sites^^}](https://${sites}.${COMPOSE_PROJECT_NAME,,}.local) :: ${typefile^^}" >> ${COMPOSE_PROJECT_NAME,,}.md
+      docker restart homelab-webserver
+      ln
+      L1 $LIGTH_CYAN " done ... (${sites}.${COMPOSE_PROJECT_NAME,,}.local)" $WHITE "⛁" "."
+   else
+      L1 $LIGTH_CYAN " The website already exists ... (${sites}.${COMPOSE_PROJECT_NAME,,}.local)" $WHITE "⛁" "."
+   fi
 }
